Split ID* and IDC* code
