+++ src/y0/algorithm/identify/id_star.py
@@ -6,7 +6,6 @@ from typing import Collection, FrozenSet, Mapping, Optional, Set, Tuple
 
 from .cg import has_same_function, make_counterfactual_graph
 from .utils import Unidentifiable
-from ..conditional_independencies import are_d_separated
 from ...dsl import (
     CounterfactualEvent,
     CounterfactualVariable,
@@ -25,15 +24,11 @@ from ...graph import NxMixedGraph
 
 __all__ = [
     "id_star",
-    "idc_star",
 ]
 
 
-class Inconsistent(ValueError):
-    pass
-
-
 def id_star(graph: NxMixedGraph, event: CounterfactualEvent) -> Expression:
+    """Apply the ``ID*`` algorithm to the graph."""
     # Line 0
     if id_star_line_1(graph, event) is not None:
         return One()
@@ -44,17 +39,14 @@ def id_star(graph: NxMixedGraph, event: CounterfactualEvent) -> Expression:
     new_query = id_star_line_3(graph, event)
     if new_query is not None:
         return id_star(graph, new_query)
-
     # Line 4:
-
     new_graph, new_event = id_star_line_4(graph, event)
-    vertices = set(new_graph.nodes())
     # Line 5:
     if new_event == Zero():
         return new_event
     # Line 6:
-
     if not new_graph.is_connected():
+        # FIXME missing third parameter to `id_star_line_6`
         summand, interventions_of_each_district = id_star_line_6(new_graph, new_event)
         return Sum.safe(
             Product.safe(
@@ -64,13 +56,11 @@ def id_star(graph: NxMixedGraph, event: CounterfactualEvent) -> Expression:
             summand,
         )
     # Line 7:
+    elif id_star_line_8(new_graph, new_query):
+        raise Unidentifiable
     else:
-        # Line 8 is syntactically impossible with the dsl
-        if id_star_line_8(new_graph, new_query):
-            raise Unidentifiable
-        else:
-            # Line 9
-            return id_star_line_9(new_graph)
+        # Line 9
+        return id_star_line_9(new_graph)
 
 
 def id_star_line_1(graph: NxMixedGraph, event: CounterfactualEvent) -> Optional[One]:
@@ -232,75 +222,7 @@ def id_star_line_9(graph: NxMixedGraph) -> Probability:
     return P[interventions](Variable(node.name) for node in graph)
 
 
-def idc_star_line_2(graph: NxMixedGraph, query: CounterfactualEvent) -> Expression:
-    r"""Run line 2 of the IDC* algorithm.
-
-    The second line states that if :math:`\event` contains a counterfactual which violates
-    the Axiom of Effectiveness (Pearl, 2000), then :math:`\event` is inconsistent, and we
-    return probability 0.
-    """
-    delta = query.parents
-    event_and_delta = query.uncondition()
-    return make_counterfactual_graph(graph, event_and_delta)
-
-
-def idc_star_line_4(graph: NxMixedGraph, query: CounterfactualEvent) -> bool:
-    r"""Run line 4 of the IDC* algorithm.
-
-    Line 4 of IDC* is the central line of the algorithm and is
-    analogous to line 1 of IDC. In IDC, we moved a value
-    assignment :math:`Z = z` from being observed to being fixed if
-    there were no back-door paths from :math:`Z` to the outcome
-    variables :math:`Y` given the context of the effect of
-    :math:`do(\mathbf{x})`. Here in IDC*, we move a counterfactual
-    value assignment :math:`Y_\mathbf{x} = y` from being observed (that is being a
-    part of :math:`\delta`), to being fixed (that is appearing in every
-    subscript of :math:`\event'` ) if there are no back-door paths from :math:`Y_\mathbf{x}` to
-    the counterfactual of interest :math:`\event'` .
-    """
-    event = set(query.children)
-    raise NotImplementedError
-
-
-def idc_star(graph: NxMixedGraph, query: CounterfactualEvent) -> Expression:
-    r"""Run the IDC* algorithm.
-
-    INPUT:
-        G a causal diagram,
-        :math:`\event` a conjunction of counterfactual outcomes,
-        :math:`\delta` a conjunction of counterfactual observations
-    :returns: an expression for :math:`P(\event | \delta)` in terms of P, FAIL, or UNDEFINED
-    """
-    delta = set(query.parents)
-    if not delta:
-        raise ValueError(f"Query {query} must be conditional")
-    # Line 1:
-    if not id_star(graph, P(delta)):
-        raise ValueError(f"Query {query} is undefined")
-    event = set(query.children)
-    # Line 2:
-    try:
-        new_graph, new_query = make_counterfactual_graph(graph, P(event.union(delta)))
-        new_event = {g for g in event if g in new_query.children}
-        new_delta = {d for d in delta if d in new_query.children}
-        vertices = set(new_graph.nodes())
-    # Line 3:
-    except Inconsistent:
-        # (f"query {event.union(delta)} is inconsistent")
-        return Zero()
-    # Line 4:
-    for counterfactual in new_delta:
-        # TODO do we need to extend the notion of d-separation from 1-1 to 1-many?
-        if are_d_separated(new_graph.remove_out_edges(counterfactual), counterfactual, new_event):
-            counterfactual_value = Variable(counterfactual.name)
-            parents = new_delta - {counterfactual}
-            children = {g.remove_in_edges(counterfactual_value) for g in new_event}
-            return idc_star(graph, P(children | parents))
-    # Line 5:
-    estimand = id_star(graph, new_query)
-    return estimand.marginalize(vertices - delta)
-
-
+# FIXME this is unused -> delete it
 def get_interventions(query: CounterfactualEvent) -> Collection[Variable]:
     r"""Generate new Variables from the subscripts of counterfactual variables in the query."""
     interventions = set()
@@ -313,13 +235,14 @@ def get_interventions(query: CounterfactualEvent) -> Collection[Variable]:
 # TODO update the docs: is this generally applicable, or only to graphs
 #  constructed as parallel worlds? perhaps update the name?
 
-
+# FIXME this is unused -> delete it
 def has_same_domain_of_values(node1: Variable, node2: Variable) -> bool:
     if isinstance(node1, CounterfactualVariable) and isinstance(node2, CounterfactualVariable):
         treatment1, treatment2 = _get_treatment_variables(node1), _get_treatment_variables(node2)
     raise NotImplementedError
 
 
+# FIXME this is unused -> delete it
 def has_same_value(event: Collection[Variable], node1: Variable, node2: Variable) -> bool:
     n1 = None
     for node in event:
